chore: 파일분리, 그룹핑 (#4)
