+++ SwiftUI-SimpleSocial/SwiftUI-SimpleSocial/Logic/Store.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+// MARK: - Store ObservableObject
+class Store: ObservableObject {
+    @Published var graph: Dictionary<UserID, User> = [:]
+    
+    init() {
+        self.graph = DummyGenerator().make()
+    }
+}
